cc-ts/agent_kernel/requirements.txt
anthropic>=0.25.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
